Adding MTA assignment - The subway system
The system is designed to take the starting station and the destintion.
Then it will find the first path it detects so may not be the most optimal

There was no time for me to do the path of least resistance for this solution. ( next version..).

The system can accept any combination and it will plot a course as the outcome.

I didn't have time to inform that you need to change at a junction station, (next versio)

It has been designed such that you can load any subway map and get an outcome.

---------------Sample output-----------------
eg:

Planning your trip from 1st  to  Time Square
========= Subway Journey Planner ============

1st
3rd
Union Central
23rd
28th
34th

========= Thank you for using the Journey Planner !!!! =========